Add coordinators
The goddamn code won't compile because of what I believe to be a bug in
XCode. XCode will not let me pass an object in an `NSNotification`
without first casting it to `AnyObject`. Of course, a forced, up-casting
to `AnyObject` fails and an implicit up-casting will result in `nil`.

Here's the error message:

> Cannot convert value of type 'Keno' to expected argument type
> 'AnyObject?'

What the fuck.