Give up on Swish
Either Swish or the framework it relies on, Result, is incompatible with
Swift 2.2 and, therefore, will not compile. I waited 1 week for a fix;
nothing came.

I originally used Swish because I thought it was simpler than Alamofire.
While the LOC may be fewer in Swish, it's still harder to use and has
far less momentum than Alamofire.